+++ src/bpp/system.py
@@ -47,6 +47,9 @@ from bpp.models import (
     Wydawnictwo_Ciagle_Zewnetrzna_Baza_Danych,
     Autor_Dyscyplina,
     Wydawnictwo_Zwarte_Zewnetrzna_Baza_Danych,
+    Grant,
+    Grant_Rekordu,
+    Element_Repozytorium,
 )
 from bpp.models.const import GR_WPROWADZANIE_DANYCH
 from bpp.models.konferencja import Konferencja
@@ -91,6 +94,7 @@ groups = {
         Rodzaj_Prawa_Patentowego,
         Dyscyplina_Naukowa,
         Zewnetrzna_Baza_Danych,
+        Grant,
     ],
     "struktura": [Uczelnia, Wydzial, Jednostka, Jednostka_Wydzial],
     GR_WPROWADZANIE_DANYCH: [
@@ -116,6 +120,8 @@ groups = {
         Wydawnictwo_Zwarte_Zewnetrzna_Baza_Danych,
         Wydawca,
         Poziom_Wydawcy,
+        Grant_Rekordu,
+        Element_Repozytorium,
     ],
     "indeks autorów": [Autor, Autor_Jednostka],
     "administracja": [User, Group, SearchForm],